# backend/openapi.json
{
    "openapi": "3.1.0",
    "paths": {
...
        "/users/me": {
            "get": {
                "operationId": "Users/getUser",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Users"
                ]
            },
            "patch": {
                "operationId": "Users/updateUser",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Users"
                ]
            }
        },
        "/users/me/email": {
            "patch": {
                "operationId": "Users/updateEmail",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEmailDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Users"
                ]
            }
        },
        "/users/me/phone": {
            "patch": {
                "operationId": "Users/updatePhone",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePhoneDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Users"
                ]
            }
        },
        "/users/me/icon": {
            "patch": {
                "operationId": "Users/updateIcon",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateIconDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Users"
                ]
            }
        },
        "/users": {
            "get": {
                "operationId": "Admin/getUsers",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Admin"
                ]
            }
        }
    },
    "info": {
        "title": "API",
        "description": "",
        "version": "1.0.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "ApiError": {
                "type": "object",
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ErrorCode"
                    },
                    "status": {
                        "type": "number"
                    },
                    "message": {
                        "type": "string"
                    },
                    "detail": {
                        "type": "object"
                    }
                },
                "required": [
                    "code",
                    "status",
                    "message"
                ]
            },
            "ApiResponse": {
                "type": "object",
                "properties": {
                    "version": {
                        "type": "string"
                    },
                    "endpoint": {
                        "type": "string"
                    },
                    "docs": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    },
                    "error": {
                        "$ref": "#/components/schemas/ApiError"
                    }
                },
                "required": [
                    "version",
                    "endpoint",
                    "docs",
                    "timestamp"
                ]
            },
...
            "ErrorCode": {
                "type": "string",
                "oneOf": [
                    {
                        "title": "GenericError",
                        "const": "generic_error"
                    },
                    {
                        "title": "InvalidRequest",
                        "const": "invalid_request"
                    },
                    {
                        "title": "ValidationError",
                        "const": "validation_error"
                    },
                    {
                        "title": "Unauthorized",
                        "const": "unauthorized"
                    },
                    {
                        "title": "Forbidden",
                        "const": "forbidden"
                    },
                    {
                        "title": "NotFound",
                        "const": "not_found"
                    },
                    {
                        "title": "EmailTaken",
                        "const": "email_taken"
                    },
                    {
                        "title": "ConfigurationError",
                        "const": "configuration_error"
                    }
                ]
            },
            "Role": {
                "type": "string",
                "oneOf": [
                    {
                        "title": "Admin",
                        "const": "admin"
                    }
                ]
            },
...
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "birthdate": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "emailVerified": {
                        "type": "boolean"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "phoneVerified": {
                        "type": "boolean"
                    },
                    "locale": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    }
                },
                "required": [
                    "id",
                    "name",
                    "surname",
                    "birthdate",
                    "email",
                    "emailVerified",
                    "phone",
                    "phoneVerified",
                    "locale",
                    "timezone",
                    "icon",
                    "roles"
                ]
            },
...
            "UpdateEmailDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    }
                },
                "required": [
                    "email"
                ]
            },
            "UpdateIconDto": {
                "type": "object",
                "properties": {
                    "icon": {
                        "type": "string"
                    }
                }
            },
            "UpdatePhoneDto": {
                "type": "object",
                "properties": {
                    "phone": {
                        "type": "string"
                    }
                }
            },
            "UpdateUserDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "birthdate": {
                        "type": "string"
                    },
                    "locale": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    }
                },
                "required": [
                    "name"
                ]
            },
...
        }
    }
}